Methods were organized within header files for better readability
